Extracting ParameterExtractor's from main class
